Profile UI fixes + Cancellation Feature

+++ backend/src/main/java/com/team26/backend/model/City.java
@@ -1,9 +1,13 @@
 package com.team26.backend.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
-
 @Getter
 @Setter
 @Entity
@@ -12,12 +16,12 @@ public class City {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "cityid")   // map to existing column
-    private Long cityId;
+    @Column(name = "cityid")
+    private Integer cityId; // Changed from Long to Integer
 
     @Column(name = "name")
     private String name;
 
-    @Column(name = "isinternational")  // map to existing column
+    @Column(name = "isinternational")
     private Boolean isInternational;
 }
