+++ backend/src/main/java/com/team26/backend/model/User.java
@@ -1,9 +1,13 @@
 package com.team26.backend.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
-
 @Getter
 @Setter
 @Entity
@@ -13,7 +17,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "userid")
-    private Long userId;
+    private Integer userId; // Changed from Long to Integer
 
     @Column(name = "username", unique = true)
     private String username;
@@ -24,6 +28,4 @@ public class User {
     private String password;
 
     private String role = "PASSENGER";
-
-    // getters & setters
 }
